Player Rendering
- Added the basics of player rendering, they do not animate yet.
- Names do not appear properly centered yet.
- Fixed an issue where I assumed a looped count would always be a
dictionary entry.
- Players will not join/leave the appropriate maps when they join the
game or disconnect.

+++ Extensions/Classes/Player.cs
@@ -53,8 +53,8 @@ namespace Extensions.Database {
 
         // Position
         public Int32            Map             { get; set; }
-        public Byte             X               { get; set; }
-        public Byte             Y               { get; set; }
+        public Int32            X               { get; set; }
+        public Int32            Y               { get; set; }
         public Byte             Direction       { get; set; }
         #endregion
